BIGTOP-3102: docker-hadoop.sh WARNING: The scale command is deprecated
When running docker-hadoop.sh to create a virtual Hadoop cluster on top
of Docker containers, there is a warning message:

WARNING: The scale command is deprecated. Use the up command with the
--scale flag instead.

Checking the shell file, the warning is spit out by docker-compose.
There is a suggestion from docker-compose official website to modify
it using 'up':

https://docs.docker.com/compose/reference/scale/

This patch made this update. And verified using Arm aarch64.

Signed-off-by: Guodong Xu <guodong.xu@linaro.org>